fix(send-and-archive): Don't validate defaultSendAction key in config (#3031)
Before this commit, opening a new composer window would reset the
default send action to just `send`, even if you had selected it to be
`send-and-archive` via preferences.

This happened because on window initialization we run a config schema
validation, and coerce any values that are not valid under the schema
(e.g. make sure that the value for a setting is one of possible values specified
in the enum). Given that the schema is static, but the values for the enum for
this particular setting are dynamically calculated when we render the selector,
it doesn't make sense to validate the values for this setting.

To prevent validation, we just need to remove the `enum`, and `enumLabels`
from the static config-schema.