[search-index] Add isSearchIndexed field to searchable models
Summary:
The isModelIndexed function was extremely slow because it was scanning
the unindexed 'content_id' field in the search index table. This diff
adds a field to the searchable models (currently Thread, Contact, and
Event) that stores whether that particular model has been indexed or
not, allowing us to avoid doing the O(n) scan of the search index table.

Test Plan: Run locally, verify that isModelIndexed is no longer a bottleneck

Reviewers: juan, evan

Reviewed By: evan

Differential Revision: https://phab.nylas.com/D3695